backend/requirements.txt
django
psycopg2
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-cors-headers
requests
requests-mock